Small refactoring so that types and extensions used by the parser are combined in one file.
